Bugfix: Hotbackup may lost data.
PROBLEM
=======
During the backup process with SmartEngine, it's necessary to obtain
a consistent checkpoint. The expectation is that the final backup data
will remain consistent with this checkpoint. There is a flaw in the
logic of obtaining the consistency checkpoint, which might lead to
the deletion of WAL logs contained within this consistency checkpoint
before they are copied. This could result in the loss of data corresponding
to those WAL logs.

```
int BackupSnapshotImpl::acquire_snapshots(DB *db)
{
  if (FAILED(db->create_backup_snapshot(...)
  //The purge of wal files may happen here.
  if (FAILED(link_files(...)))
}
```

SOLUTION
========
The entire process of acquire_snapshots disables WAL purge.