optimize: optimize columnar format storage.
Optimize the statistics collection for columnar storage format to
accurately display the original data size in column format.

Improve the extent filling algorithm to enhance the extent fill rate
and reduce storage costs.The specific improvement approach is to use
the compressed data block and index block when calculating whether
the extent meets the conditions for switching.