feat: persistent snapshot multi version (#32)
fix #20

## Description

The MySQL Leader role node periodically generates snapshot and persists
it to the object store, and when a Leader role change occurs, there may
be two Leader nodes at the same time (the old Leader did not exit in
time, the new Leader has been replaced) that both generate snapshot and
persist to the object store, a situation occurs where snapshot in the
object store is overwritten by two nodes over each other.

## Solution

Append a version number to each snapshot name in the object store, using
the term in the Raft protocol, to ensure that the snapshot generated
when the Leader role switches between different nodes in the Raft is
unique between the nodes.
When an instance is restarted or a snapshot is recovered, the latest
term's snapshot index file shall prevail. Other lower version snapshots
will be recycled.
At the same time, when designing multi-version snapshots, it must also
be compatible with single-version snapshots.
When there is only a single-version snapshot, the single-version
snapshot is used. Otherwise, when there are multi-version snapshots, the
single-version snapshot will be recycled as an expired snapshot.

The current data node raft term is 1 and the 10 snapshots
generated,persistent single-version snapshot format : ` snapshot.index`
```
snapshot.index
innodb_000010/
smartengine_000010/
```

The current data node raft term is 1 and the 10 snapshots generated,
persistent multi-version snapshot format : ` snapshot.{term}.index`

```
snapshot.000001.index
innodb_000001_000010/
smartengine_000001_000010/
```